requirements.txt
Flask-HTTPAuth==3.3.0
google==2.0.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
SQLAlchemy==1.3.5
Werkzeug==0.15.5

